requirements.txt
requests
ntplib
rpi_ws281x
waveshare-epaper
pydub
